renamed OldHubIOStream to TcpIOStream to reflect its usability with programs such as ser2net.
refactored factory method in IOStream to allow tcp as a prefix within
config
