+++ bundles/binding/org.openhab.binding.insteonplm/src/main/java/org/openhab/binding/insteonplm/internal/driver/IOStream.java
@@ -13,7 +13,6 @@ import java.io.InputStream;
 import java.io.OutputStream;
 
 import org.openhab.binding.insteonplm.internal.driver.hub.HubIOStream;
-import org.openhab.binding.insteonplm.internal.driver.hub.OldHubIOStream;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -33,7 +32,7 @@ public abstract class IOStream {
 
     /**
      * read data from iostream
-     * 
+     *
      * @param b byte array (output)
      * @param offset offset for placement into byte array
      * @param readSize size to read
@@ -60,7 +59,7 @@ public abstract class IOStream {
 
     /**
      * Write data to iostream
-     * 
+     *
      * @param b byte array to write
      */
     public void write(byte[] b) {
@@ -81,7 +80,7 @@ public abstract class IOStream {
 
     /**
      * Opens the IOStream
-     * 
+     *
      * @return true if open was successful, false if not
      */
     public abstract boolean open();
@@ -93,7 +92,7 @@ public abstract class IOStream {
 
     /**
      * reconnects the stream
-     * 
+     *
      * @return true if reconnect succeeded
      */
     private synchronized boolean reconnect() {
@@ -103,13 +102,13 @@ public abstract class IOStream {
 
     /**
      * Creates an IOStream from an allowed config string:
-     * 
+     *
      * /dev/ttyXYZ (serial port like e.g. usb: /dev/ttyUSB0 or alias /dev/insteon)
-     * 
+     *
      * /hub2/user:password@myinsteonhub.mydomain.com:25105,poll_time=1000 (insteon hub2 (2014))
-     * 
+     *
      * /hub/myinsteonhub.mydomain.com:9761
-     * 
+     *
      * @param config
      * @return reference to IOStream
      */
@@ -117,7 +116,7 @@ public abstract class IOStream {
     public static IOStream s_create(String config) {
         if (config.startsWith("/hub2/")) {
             return makeHub2014Stream(config);
-        } else if (config.startsWith("/hub/")) {
+        } else if (config.startsWith("/hub/") || config.startsWith("/tcp/")) {
             return makeOldHubStream(config);
         } else {
             return new SerialIOStream(config);
@@ -153,12 +152,12 @@ public abstract class IOStream {
         return new HubIOStream(hp.host, hp.port, pollTime, user, pass);
     }
 
-    private static OldHubIOStream makeOldHubStream(String config) {
+    private static TcpIOStream makeOldHubStream(String config) {
         config = config.substring(5); // Get rid of the /hub/ part
         String[] parts = config.split(","); // split off options at the end, if any
         String[] hostPort = parts[0].split(":");
         HostPort hp = new HostPort(hostPort, 9761);
-        return new OldHubIOStream(hp.host, hp.port);
+        return new TcpIOStream(hp.host, hp.port);
     }
 
     private static class HostPort {
